Suppress warnings from reading sbml model.
There are hundreds of the same warning issued
when reading in the FBA model from the xml files.
These are clutering up the notebooks and so are
now suppressed. The source of the warnings were
the wrong case for 'and'/'or', which were being
read correctly and missing values of the charge
of some of the metabolites, which are not used.